Aborting wp-install if WordPress is already installed

+++ src/Bootstrap.php
@@ -85,6 +85,11 @@ class Bootstrap
         $this->log->addDebug('Running Bootstrap::install');
         $wpcmd = $this->utils->getWpCommand();
 
+        if (file_exists($this->localSettings->dbhost . '/wp-config.php')) {
+            $this->log->addInfo('WordPress already installed');
+            return;
+        }
+
         $cmd = 'rm -rf ~/.wp-cli/cache/core/';
         $this->utils->exec($cmd);
